Added Module 6

+++ taskManager/src/main/java/com/fourimpact/taskManager/entity/Category.java
@@ -28,7 +28,7 @@ public class Category {
         this.description = description;
     }
 
-    public int getId() { return id; }
+    public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
